+++ src/gufo/thor/config.py
@@ -53,6 +53,7 @@ class NocConfig(object):
         installation_name: The installation name which will be shown
             in the interface.
         theme: Web interface theme. One of: `noc`, `gray`.
+        language: Web interface language.
         migrate: Run migrations on start
         config: User-defined config.
     """
@@ -62,6 +63,7 @@ class NocConfig(object):
     custom: Optional[str] = None
     installation_name: str = "Unconfigured Installation"
     theme: Literal["noc", "gray"] = "noc"
+    language: Literal["en", "ru"] = "en"
     migrate: bool = True
     config: Optional[Dict[str, Any]] = None