+++ src/gufo/thor/config.py
@@ -47,7 +47,9 @@ class NocConfig(object):
     Attributes:
         tag: NOC image tag
         path: An optional path to the NOC source code. If not empty,
-            image's `/opt/noc`vwill be replaced with path
+            image's `/opt/noc` will be replaced with.
+        ui_path: An optional path to the compiled UI. If not empty,
+            static image's `/www` will be replaced with.
         custom: Optional path to the custom, will be mounted
             in the `/opt/noc_custom` directory.
         installation_name: The installation name which will be shown
@@ -60,6 +62,7 @@ class NocConfig(object):
 
     tag: str = "master"
     path: Optional[str] = None
+    ui_path: Optional[str] = None
     custom: Optional[str] = None
     installation_name: str = "Unconfigured Installation"
     theme: Literal["noc", "gray"] = "noc"