+++ CHANGELOG.md
@@ -16,6 +16,10 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 * migrate container no longer pollutes environment of other containers
 
+## Added
+
+* `noc.language` setting.
+
 ## Changed
 
 * Configs are deployed using docker's `configs` section. 
@@ -34,6 +38,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 * Jinja2 3.1.6
 * Ruff 0.14
 * devcontainer: Python debugger settings.
+* Use NOC images from ghcr.io/gufolabs/noc
 
 ## 0.11.0 - 2025-09-12